No misbinding and fixed scaffold enum (messy). Also switched step to long long int.

+++ include/annealing_simulation.h
@@ -16,11 +16,11 @@ namespace Annealing {
                     InputParameters& params);
             void run();
         private:
-            void update_internal(long int) {};
+            void update_internal(long long int) {};
             double m_max_temp;
             double m_min_temp;
             double m_temp_interval;
-            int m_steps_per_temp;
+            long long int m_steps_per_temp;
     };
 }
